Different Counters for different exercises :)

+++ WorkoutTracker/SwiftUIViews/WorkoutSelectView.swift
@@ -10,14 +10,14 @@ import SwiftUI
 struct Exercise {
     var name: String
     var reps: Int
-    var time: Int
+    var time: Double
     var rest: Int
 }
 
 let Workout = [
-    Exercise(name: "Pull-up", reps: 10, time: 45, rest: 45),
-    Exercise(name: "Push-up", reps: 10, time: 30, rest: 45),
-    Exercise(name: "Sit-up", reps: 20, time: 60, rest: 45),
+    Exercise(name: "Pull-up", reps: 10, time: 45.0, rest: 45),
+    Exercise(name: "Push-up", reps: 10, time: 30.0, rest: 45),
+    Exercise(name: "Sit-up", reps: 20, time: 60.0, rest: 45),
 ]
 
 struct DetailView : View {
